+++ functions/theme-setup.php
@@ -15,6 +15,9 @@ function theme_setup() {
 	add_theme_support( 'wp-block-styles' );
 	add_theme_support( 'align-wide' );
 	add_theme_support( 'responsive-embeds' );
+	add_theme_support( 'custom-logo' );
+	add_theme_support( 'unlink-homepage-logo' );
+	add_theme_support( 'block-template-parts' );
 	remove_theme_support( 'core-block-patterns' );
 }
 add_action( 'after_setup_theme', __NAMESPACE__ . '\\theme_setup', 15 );