+++ index.php
@@ -1,0 +1,13 @@
+<?php
+/**
+ * The index file for the theme.
+ */
+
+?>
+<?php get_header(); ?>
+<div class="template-content">
+	<main id="skip-link-target" class="contained">
+		<?php the_content(); ?>
+	</main>
+</div>
+<?php get_footer(); ?>